+++ JimHalpert.Domain/Entity/TarefaItem.cs
@@ -8,7 +8,7 @@ namespace JimHalpert.Domain.Entity
     {
         public int TarefaItemId { get; set; }
         public int TarefaId { get; set; }
-        public byte Ordem { get; set; }
+        public int Ordem { get; set; }
         public string Descricao { get; set; }
         public int ServicoId { get; set; }
         public int HorasOrcadas { get; set; }
